requirements.txt
python-telegram-bot==13.7
alpha-vantage==2.3.1
python-dotenv==0.19.0
pandas==1.3.3
